+++ video_xblock/video_xblock.py
@@ -151,24 +151,9 @@ class VideoXBlock(
         help=_(
             'Default transcripts are uploaded automatically from a video platform '
             'to the list of available transcripts.<br/>'
-            '<b>Note: "Video API Token" should be given in order to make auto fetching possible.</b>'
+            '<b>Note: valid "Video API Token" should be given in order to make auto fetching possible.</b><br/>'
+            'Advice: disable transcripts displaying on your video service to avoid transcripts overlapping.'
         ),
-        resettable_editor=False
-    )
-
-    threeplaymedia_apikey = String(
-        default='default',
-        display_name=_('API Key'),
-        help=_('You can generate a client token following official documentation of your video platform\'s API.'),
-        scope=Scope.content,
-        resettable_editor=False
-    )
-
-    threeplaymedia_file_id = String(
-        default='default',
-        display_name=_('File Id'),
-        help=_('3playmedia file id for download bind transcripts.'),
-        scope=Scope.content,
         resettable_editor=False
     )
 
@@ -317,7 +302,7 @@ class VideoXBlock(
                 display_name=self.display_name,
                 usage_id=self.usage_id,
                 handout=self.handout,
-                transcripts=self.route_transcripts(self.transcripts),
+                transcripts=self.route_transcripts(),
                 download_transcript_allowed=self.download_transcript_allowed,
                 download_video_url=self.get_download_video_url(),
                 handout_file_name=self.get_file_name_from_path(self.handout),
@@ -351,6 +336,7 @@ class VideoXBlock(
             default_transcripts, transcripts_autoupload_message = player.get_default_transcripts(**kwargs)
         except ApiClientError:
             default_transcripts, transcripts_autoupload_message = [], _('Failed to fetch default transcripts.')
+        log.debug("Autofetch message: '{}'".format(transcripts_autoupload_message))
         # Default transcripts should contain transcripts of distinct languages only
         distinct_default_transcripts = player.clean_default_transcripts(default_transcripts)
         # Needed for frontend
@@ -380,7 +366,7 @@ class VideoXBlock(
         # whilst for Wistia, a sample authorised request is to be made to ensure authentication succeeded,
         # since it is needed for the auth status message generation and the player's state update with auth status.
         if self.token:
-            _auth_data, auth_error_message = self.authenticate_video_api(self.token)
+            _auth_data, auth_error_message = self.authenticate_video_api(self.token.encode(encoding='utf-8'))
 
         initial_default_transcripts, transcripts_autoupload_message = self._update_default_transcripts(
             player, transcripts
@@ -389,7 +375,7 @@ class VideoXBlock(
         # Prepare basic_fields and advanced_fields for them to be rendered
         basic_fields = self.prepare_studio_editor_fields(player.basic_fields)
         advanced_fields = self.prepare_studio_editor_fields(player.advanced_fields)
-        log.debug("Fetched default transcripts: {}".format(self.default_transcripts))
+        log.debug("Fetched default transcripts: {}".format(initial_default_transcripts))
         context = {
             'advanced_fields': advanced_fields,
             'auth_error_message': auth_error_message,
@@ -401,6 +387,9 @@ class VideoXBlock(
             'sources': TranscriptSource.to_dict().items(),
             # transcripts context:
             'transcripts': transcripts,
+            'transcripts_fields': self.prepare_studio_editor_fields(player.trans_fields),
+            'three_pm_fields': self.prepare_studio_editor_fields(player.three_pm_fields),
+            'transcripts_type': '3PM' if self.threeplaymedia_streaming else 'manual',
             'default_transcripts': self.default_transcripts,
             'enabled_default_transcripts': filter_transcripts_by_source(transcripts),
             'initial_default_transcripts': initial_default_transcripts,
@@ -412,6 +401,7 @@ class VideoXBlock(
         fragment.add_css(resource_string("static/css/student-view.css"))
         fragment.add_css(resource_string("static/css/transcripts-upload.css"))
         fragment.add_css(resource_string("static/css/studio-edit.css"))
+        fragment.add_css(resource_string("static/css/studio-edit-accordion.css"))
         fragment.add_javascript(resource_string("static/js/runtime-handlers.js"))
         fragment.add_javascript(resource_string("static/js/studio-edit/utils.js"))
         fragment.add_javascript(resource_string("static/js/studio-edit/studio-edit.js"))
@@ -435,7 +425,7 @@ class VideoXBlock(
         save_state_url = self.runtime.handler_url(self, 'save_player_state')
         transcripts = render_resource(
             'static/html/transcripts.html',
-            transcripts=self.route_transcripts(self.transcripts)
+            transcripts=self.route_transcripts()
         ).strip()
         return player.get_player_html(
             url=self.href, account_id=self.account_id, player_id=self.player_id,
@@ -559,7 +549,7 @@ class VideoXBlock(
                 'has_list_values': False,
                 'type': 'string',
             }
-        elif field_name in ('handout', 'transcripts', 'default_transcripts', 'token', 'threeplaymedia_apikey'):
+        elif field_name in ('handout', 'transcripts', 'default_transcripts', 'token'):
             info = self.initialize_studio_field_info(field_name, field, field_type=field_name)
         else:
             info = self.initialize_studio_field_info(field_name, field)
@@ -782,3 +772,14 @@ class VideoXBlock(
         }
         log.debug("Uploaded default transcript: {}".format(response))
         return response
+
+    def get_enabled_transcripts(self):
+        """
+        Get transcripts from different sources depending on current usage mode.
+        """
+        if self.threeplaymedia_streaming:
+            transcripts = list(self.fetch_available_3pm_transcripts())
+        else:
+            transcripts = json.loads(self.transcripts) if self.transcripts else []
+
+        return transcripts